// web/src/lib/api/generated/models/internal-products-stock-history-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * POS Kasir API
 * POS Kasir API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface InternalProductsStockHistoryResponse {
    'change_amount'?: number;
    'change_type'?: string;
    'created_at'?: string;
    'created_by'?: string;
    'current_stock'?: number;
    'id'?: string;
    'note'?: string;
    'previous_stock'?: number;
    'product_id'?: string;
    'reference_id'?: string;
}